+++ CHANGELOG.rst
@@ -5,9 +5,13 @@ History
 0.17.0 (UNRELEASED)
 ===================
 
+* Disable internal hdf file locks.
+
 * Added ``esp_thermal_efficiency_model`` attribute to ``PumpEquipmentDescription``.
 
-* Disable internal hdf file locks.
+* Added ``esp_manufacturer`` and ``esp_model`` attributes to ``PumpEquipmentDescription``.
+
+* Removed ``user_defined_esp_table`` and ``esp_parameters`` from ``PumpEquipmentDescription``.
 
 
 0.16.0 (2023-02-09)